requirements.txt
langchain==0.1.12
langchain-core==0.1.31
playwright==1.42.0
python-dotenv==1.0.1 